+++ src/dragLabel.cpp
@@ -3,7 +3,6 @@
 #include <QtGui/QMouseEvent>
 #include <iostream>
 #include <QtCore/QBasicTimer>
-#include "MainWindow.h"
 #include "dragLabel.h"
 #include "mrichtextedit.h"
 #include "dragLabelInput.h"