+++ src/main.cpp
@@ -7,13 +7,17 @@
 
 int main(int argc, char *argv[])
 {
+	std::string startup = "assets/default.vsim";
+	if (argc >= 2) {
+		startup = argv[1];
+	}
 	QApplication a(argc, argv);
-	MainWindow window(NULL);
+	MainWindow window;
 	a.addLibraryPath("plugins");
 
 	window.show();
 	VSimApp vsim(&window);
 
-	vsim.openVSim("assets/default.vsim");
+	vsim.openVSim(startup);
 	return a.exec();
 }